MGMT-16036: add unique OSImages validation rule
Add API validation for duplicate OSImages checking that each
openshiftVersion+cpuArchitecture is unique. Unless the length of the
fields is limited, the cost estimation for the CEL expression is too
high. Set the maxLength of both to a reasonable setting and set the
maxItems of the OSImages slice to a high limit.

Signed-off-by: Brady Pratt <bpratt@redhat.com>